structure for queryparsing in place

+++ src/parser/Scanner.h
@@ -1,3 +1,18 @@
+/* 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 //
 // Created by Peter G. Jensen on 10/25/18.
 //
@@ -11,9 +26,9 @@
 
 #endif
 
-#include "generated_parser.hh"
+#include "generated_QueryParser.hh"
 #include "location.hh"
-#include "Builder.h"
+#include "QueryBuilder.h"
 
 namespace mpls2pda {
 
@@ -30,7 +45,7 @@ namespace mpls2pda {
         int yylex();
 
         double last_float;
-        int last_int;
+        int64_t last_int;
         std::string last_string;
     };
